# allow faster bootstrap with less forced reflow and dom manipulation