--- a/docs/assets/owlcarousel/owl.carousel.js
+++ b/docs/assets/owlcarousel/owl.carousel.js
@@ -215,6 +215,7 @@
 		fallbackEasing: 'swing',
 
 		info: false,
+		checkVisibility: true,
 
 		nestedItemSelector: false,
 		itemElement: 'div',