Hey,

we love the OwlCarousel and have it heavily at use in production. On some pages up to 4 carousels at the same time.

The problem: On old android devices (like Galaxy S4/S5ish), DOM manipulation and forced reflows are very slow. And it get worse the more complex the page becomes.

With this PR we were able to reduce the render time of a page with 4 carousels from 10 seconds to 2,5 seconds on a Galaxy S4.

In this plunkr the difference is not that severe as its just a blank page and the items are simple text, but you already see the impact:

### default
plunkr: http://plnkr.co/edit/c52107vIiWWNJDYPKJqV?p=preview
4 carousel ~ 100ms
![image](https://cloud.githubusercontent.com/assets/905328/24912638/ff852444-1ece-11e7-9d5a-5ee18e9a6707.png)

### fast
plunkr: http://plnkr.co/edit/CSRUwANcqOgD4S7LAmDm?p=preview
4 carousel ~ 55ms
![image](https://cloud.githubusercontent.com/assets/905328/24913378/4b077352-1ed1-11e7-919f-d5e578721f51.png)

## Trick 1

Instead of
```twig
      <div class="owl-carousel owl-theme carousel-default">
        <div class="item bg-navy">item</div>
      </div>
```

we do
```twig
      <div class="owl-carousel owl-theme carousel-fast owl-loaded owl-drag">
        <div class="owl-stage-outer">
          <div class="owl-stage">
            <div class="owl-item">
              <div class="item bg-navy">item</div>
            </div>
          </div>
        </div>
      </div>
```

So in the initialize method the OwlCarousel does not need to move the DOM around anymore. The simple way still works! But you may provide the final DOM and this PR will skip that part.

## Trick 2
We call OwlCarousel with `{checkVisibility: false}` that tells it, that the carousel is always visible and just skip all `is(':visible')` checks
```javascript
$('.carousel').owlCarousel({checkVisibility: false});
```

Thats gets rid of some performance punishing force reflows.



What do you think?