@mtomov 2017-06-29
:+1:  Thanks for that. I've certainly initial load heaviness on a few sites that do 3,4,5 carousels and more, especially if placed in tabs ..